minor event and I/O code reorg
- remove unused argument to `process_events`
- a bit of file renaming and reorg; remove some dead code
- this allows code in libuv.jl to use locks and other things in Threads
